+++ src/EmuEEPROM.h
@@ -69,7 +69,7 @@ class EmuEEPROM
         virtual bool     write32(uint32_t address, uint32_t data) = 0;
         virtual bool     read16(uint32_t address, uint16_t& data) = 0;
         virtual bool     read32(uint32_t address, uint32_t& data) = 0;
-        virtual size_t   pageSize()                               = 0;
+        virtual uint32_t pageSize()                               = 0;
     };
 
     EmuEEPROM(StorageAccess& storageAccess, bool useFactoryPage)
